refactor(sonar): drop redundant (ulong) casts in trace calls (S1905)
sessionId and slotId constructor parameters are already ulong, so the (ulong) casts on the
Log.SessionTrace / Log.SlotTrace arguments are no-ops. Remove them.

+++ src/KerckhoffsLabs.Security.Cryptography.Pkcs11/Internal/Pkcs11Session.cs
@@ -250,7 +250,7 @@ internal sealed partial class Pkcs11Session : IDisposable
     /// <param name="sessionId">PKCS#11 handle of session</param>
     internal Pkcs11Session(ILowLevelPkcs11Library pkcs11Library, ulong sessionId)
     {
-        Log.SessionTrace(_logger, (ulong)sessionId, "ctor");
+        Log.SessionTrace(_logger, sessionId, "ctor");
 
         ArgumentNullException.ThrowIfNull(pkcs11Library);
